+++ src/frontend/apps/impress/src/features/docs/docs-grid/components/DocsGridActions.tsx
@@ -1,7 +1,8 @@
-import { useModal } from '@gouvfr-lasuite/cunningham-react';
+import { Button } from '@gouvfr-lasuite/cunningham-react';
+import { DropdownMenu, DropdownMenuItem } from '@gouvfr-lasuite/ui-kit';
 import dynamic from 'next/dynamic';
+import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { css } from 'styled-components';
 
 import ContentCopySVG from '@/assets/icons/ui-kit/content_copy.svg';
 import DeleteSVG from '@/assets/icons/ui-kit/delete.svg';
@@ -9,8 +10,8 @@ import DocMoveInSVG from '@/assets/icons/ui-kit/doc-move-in.svg';
 import GroupSVG from '@/assets/icons/ui-kit/group.svg';
 import KeepSVG from '@/assets/icons/ui-kit/keep.svg';
 import KeepOffSVG from '@/assets/icons/ui-kit/keep_off.svg';
-import MoreHorizSVG from '@/assets/icons/ui-kit/more_horiz.svg';
-import { DropdownMenu, DropdownMenuOption, Icon } from '@/components';
+import LeaveSVG from '@/assets/icons/ui-kit/leave.svg';
+import MoreSVG from '@/assets/icons/ui-kit/more_horiz.svg';
 import {
   Doc,
   KEY_LIST_DOC,
@@ -41,17 +42,28 @@ const ModalRemoveDoc = dynamic(
   { ssr: false },
 );
 
+const ConfirmationLeaveModal = dynamic(
+  () =>
+    import('@/docs/doc-share/components/ConfirmationLeaveModal').then(
+      (mod) => ({
+        default: mod.ConfirmationLeaveModal,
+      }),
+    ),
+  { ssr: false },
+);
+
 interface DocsGridActionsProps {
   doc: Doc;
 }
 
 export const DocsGridActions = ({ doc }: DocsGridActionsProps) => {
   const { t } = useTranslation();
-  const restoreFocus = useFocusStore((state) => state.restoreFocus);
-
-  const deleteModal = useModal();
-  const shareModal = useModal();
-  const importModal = useModal();
+  const { restoreFocus, addLastFocus } = useFocusStore();
+  const [openDropdown, setOpenDropdown] = useState(false);
+  const [isModalRemoveOpen, setIsModalRemoveOpen] = useState(false);
+  const [isModalLeaveOpen, setIsModalLeaveOpen] = useState(false);
+  const [isModalShareOpen, setIsModalShareOpen] = useState(false);
+  const [isModalMoveOpen, setIsModalMoveOpen] = useState(false);
   const { untitledDocument } = useTrans();
 
   const { mutate: duplicateDoc } = useDuplicateDoc({
@@ -69,7 +81,7 @@ export const DocsGridActions = ({ doc }: DocsGridActionsProps) => {
     listInvalidQueries: [KEY_LIST_DOC, KEY_LIST_FAVORITE_DOC],
   });
 
-  const options: DropdownMenuOption[] = [
+  const options: DropdownMenuItem[] = [
     {
       label: doc.is_favorite ? t('Unpin') : t('Pin'),
       icon: doc.is_favorite ? (
@@ -91,7 +103,7 @@ export const DocsGridActions = ({ doc }: DocsGridActionsProps) => {
       label: t('Share'),
       icon: <GroupSVG width={24} height={24} aria-hidden="true" />,
       callback: () => {
-        shareModal.open();
+        setIsModalShareOpen(true);
       },
 
       testId: `docs-grid-actions-share-${doc.id}`,
@@ -100,15 +112,15 @@ export const DocsGridActions = ({ doc }: DocsGridActionsProps) => {
       label: t('Move into a doc'),
       icon: <DocMoveInSVG width={24} height={24} aria-hidden="true" />,
       callback: () => {
-        importModal.open();
+        setIsModalMoveOpen(true);
       },
-      testId: `docs-grid-actions-import-${doc.id}`,
-      show: doc.abilities.move,
+      testId: `docs-grid-actions-move-${doc.id}`,
+      isHidden: !doc.abilities.move,
     },
     {
       label: t('Duplicate'),
       icon: <ContentCopySVG width={24} height={24} aria-hidden="true" />,
-      disabled: !doc.abilities.duplicate,
+      isDisabled: !doc.abilities.duplicate,
       callback: () => {
         duplicateDoc({
           docId: doc.id,
@@ -119,72 +131,87 @@ export const DocsGridActions = ({ doc }: DocsGridActionsProps) => {
       showSeparator: true,
     },
     {
+      label: t('Leave'),
+      icon: <LeaveSVG width={24} height={24} aria-hidden="true" />,
+      callback: () => {
+        setIsModalLeaveOpen(true);
+      },
+    },
+    {
       label: t('Delete'),
       icon: <DeleteSVG width={24} height={24} aria-hidden="true" />,
-      callback: () => deleteModal.open(),
-      disabled: !doc.abilities.destroy,
+      callback: () => {
+        setIsModalRemoveOpen(true);
+      },
+      isHidden: !doc.abilities.destroy,
       testId: `docs-grid-actions-remove-${doc.id}`,
     },
   ];
-
-  const documentTitle = doc.title || untitledDocument;
-  const menuLabel = t('Open the menu of actions for the document: {{title}}', {
-    title: documentTitle,
-  });
 
   return (
     <>
       <DropdownMenu
         options={options}
-        label={menuLabel}
-        aria-label={t('More options')}
-        buttonCss={css`
-          &:hover {
-            background-color: unset;
-          }
-        `}
+        isOpen={openDropdown}
+        shouldCloseOnInteractOutside={() => true}
+        onOpenChange={setOpenDropdown}
       >
-        <Icon
+        <Button
           data-testid={`docs-grid-actions-button-${doc.id}`}
-          icon={<MoreHorizSVG width={16} height={16} aria-hidden="true" />}
-          $theme="brand"
-          $variation="secondary"
-          $css={css`
-            cursor: pointer;
-            &:hover {
-              opacity: 0.8;
-            }
-          `}
-          aria-hidden="true"
+          aria-label={t(
+            'Open the menu of actions for the document: {{title}}',
+            {
+              title: doc.title || untitledDocument,
+            },
+          )}
+          size="small"
+          icon={<MoreSVG width={16} height={16} aria-hidden="true" />}
+          color="neutral"
+          variant="tertiary"
+          onClick={(e) => {
+            e.stopPropagation();
+            e.preventDefault();
+            setOpenDropdown((o) => !o);
+            addLastFocus(e.currentTarget);
+          }}
         />
       </DropdownMenu>
 
-      {deleteModal.isOpen && (
+      {isModalRemoveOpen && (
         <ModalRemoveDoc
           onClose={() => {
-            deleteModal.onClose();
+            setIsModalRemoveOpen(false);
             restoreFocus();
           }}
           doc={doc}
         />
       )}
-      {shareModal.isOpen && (
+      {isModalShareOpen && (
         <DocShareModal
           doc={doc}
           onClose={() => {
-            shareModal.close();
+            setIsModalShareOpen(false);
             restoreFocus();
           }}
         />
       )}
-      {importModal.isOpen && (
+      {isModalLeaveOpen && (
+        <ConfirmationLeaveModal
+          onClose={() => {
+            setIsModalLeaveOpen(false);
+            restoreFocus();
+          }}
+          doc={doc}
+        />
+      )}
+      {isModalMoveOpen && (
         <DocMoveModal
           doc={doc}
           onClose={() => {
-            importModal.close();
+            setIsModalMoveOpen(false);
             restoreFocus();
           }}
-          isOpen={importModal.isOpen}
+          isOpen={isModalMoveOpen}
         />
       )}
     </>